Split the code into two classes
